msm: sensor: ois: Prevent accessing user space buffer directly
In msm_ois_control(), userspace buffer is directly accessed
which is violating PAN (Privilege Access Never).
So to prevent this, data which needs to be copied to user buffer is
gathered into a local buffer using copy_from_user() and at the end,
this local buffer contents are copied back to buffer supplied
from user space using copy_to_user().

Change-Id: Iad50fc305f49a4dd3d880ecce5963692de97b5db
Signed-off-by: TheStrix <parthbhatia98@gmail.com>