sched: core_ctl: Add core_ctl_notifier_register()
Add a notifier API to register for the number of big tasks and default
related thread group load. The notifications are sent when a WALT
window is rolled over. Since the window rollover happens from a IRQ
work, an atomic notifier chain is used.

Change-Id: I74ad038a475cba7282753b6d8079d7414f6d5f87
Signed-off-by: Pavankumar Kondeti <pkondeti@codeaurora.org>
Just use this to pass compilation
Signed-off-by: Ratoriku <a1063021545@gmail.com>