ARM/dts: xiaomi: sdm636: Add lower freq for little cluster
Based on the description of commit [1], while the CPU is idling it
causes a lot of power drain. So add the lowest frequency for little
clusters so that when the CPU is idle there is not much power drain.

The energy model and cost data for lowest frequency little cluster
are taken from the calculation results from the freqbench data [2].

* [1]: https://github.com/kerneltoast/android_kernel_google_wahoo/commit/fdb50fc69b1b605cafb07537c9eb3808ba305313
* [2]: https://github.com/kdrag0n/freqbench/commit/24990db

Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>