drivers: iio, power: Keep Xiaomi changes for tulip device
Changing the iio driver and power from the CAF tagged
"LA.UM.9.2.r1-02000-SDMxx0.0" causes errors on tulip
devices such as:

1. Incorrect reading of battery temperature while charging
2. Error reading current while charging

This can be fixed by restarting the device, but this is only
while. If the temperature increases, an error will automatically occur
repeated. This is because there may be a slight incompatibility of iio
and the power driver from the CAF driver with tulip devices, so keep
changes from Xiaomi by restore some changes apply only to tulip
devices, and exclude changes from CAF for tulip devices.

Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>