kernel: trace: Don't set tracer function config features as no-op
In commit df56e14 some of the trackers were disabled,
but the config of tracer function in Makefile was made
a no-op and this is a bit redundant. But if to disable config
an set it to "# CONFIG_... is not set" or "CONFIG_...=n" or
delete the config in defconfig. So remove the "#" in some
config tracer function so that it can be reactivated.

Signed-off-by: Fiqri Ardyansyah <fiqri15072019@gmail.com>